requirements.txt
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.24.0
sentence-transformers>=2.2.2
datasets>=2.15.0
peft>=0.6.0
bitsandbytes>=0.41.2.post2
faiss-gpu>=1.7.4
numpy>=1.24.0
